Comment out function call that connects to db
The app will break and "npm run serverstart" won't work upon first
starting the app because there is no "MONGODB_URL" variable set in the
.env file. By commenting out this line we enable the boilerplate
code necessary to connect the the db to stay in this template, while
also providing a better developer experience by not crashing the app
when first run before any code has been written. The only downside is
that the developer will have to be sure to uncomment the code after
storing the connection string in .env, but this should be a normal
part of the development process anyway.